Rename LoxoneConfig to LoxoneApp
LoxoneConfig was kind of confusing, LoxoneApp more suits the purpose.
