+++ src/main/java/cz/smarteon/loxone/app/AlarmControl.java
@@ -1,4 +1,4 @@
-package cz.smarteon.loxone.config;
+package cz.smarteon.loxone.app;
 
 import cz.smarteon.loxone.LoxoneUuid;
