+++ src/main/java/cz/smarteon/loxone/app/UnknownControl.java
@@ -1,4 +1,4 @@
-package cz.smarteon.loxone.config;
+package cz.smarteon.loxone.app;
 
 public final class UnknownControl extends Control {
