+++ src/main/java/cz/smarteon/loxone/system/status/MiniserverStatus.java
@@ -6,13 +6,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 import cz.smarteon.loxone.PercentDoubleDeserializer;
-import cz.smarteon.loxone.config.MiniserverType;
+import cz.smarteon.loxone.app.MiniserverType;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static java.util.Objects.requireNonNull;
