+++ src/test/groovy/cz/smarteon/loxone/app/MiniserverInfoTest.groovy
@@ -1,4 +1,4 @@
-package cz.smarteon.loxone.config
+package cz.smarteon.loxone.app
 
 import cz.smarteon.loxone.message.SerializationSupport
 import spock.lang.Specification
